feat: expand HTTP decoding, flow tracking, and diagnostics
- Add debug module to rust-shark-core and extend HTTP decoder with richer parsing support
- Enhance flow tracker, filter evaluation, and IPC protocol with additional state and message handling
- Validate --buffer-size CLI argument at parse time to reject zero values
- Improve TUI detail tree, packet list, and monitor startup logging for better runtime visibility
- Bump workspace version to 0.1.1 and normalize comment dividers to ASCII

+++ crates/rust-shark-core/src/flow/mod.rs
@@ -59,8 +59,8 @@ pub enum Dir {
 #[derive(Debug, Clone)]
 pub struct FlowState {
     // Retransmission detection (TCP only).
-    pub max_seq_low: Option<u64>,
-    pub max_seq_high: Option<u64>,
+    pub max_seq_low: Option<u32>,
+    pub max_seq_high: Option<u32>,
     pub fin_low: bool,
     pub fin_high: bool,
     pub rst_seen: bool,
